requirements.txt
langchain~=0.0.230
openai
unstructured
faiss-cpu
gradio~=3.36.1

PyYAML~=6.0
pydantic~=1.10.11